server/requirements.txt
alembic==1.13.3
APScheduler==3.10.4
async-timeout==5.0.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==2.1.5
msgspec==0.18.6
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.33
tomli==2.0.2
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.4
GitPython==3.1.43